refactor(anthropic): isolate the image normalize codec and cache (split S03 L1/3)
